+++ app/src/main/java/bonsai/habit/usageStatistic/alarm/AlarmReceiver.kt
@@ -9,6 +9,10 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.util.Log
+import bonsai.habit.usageStatistic.BonsaiStateManager
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 
 /**
  * Alarm receiver that is called when the alarm is triggered.
@@ -16,10 +20,22 @@ import android.util.Log
  */
 class AlarmReceiver: BroadcastReceiver() {
 
-    // TODO avoid long running tasks
+    companion object {
+        private const val TAG = "AlarmReceiver"
+    }
+
     override fun onReceive(context: Context?, intent: Intent?) {
         if (intent?.action == null) {
-            Log.d(this::class.java.name, "AlarmReceiver called")
+            if (context != null) {
+                Log.d(TAG, "AlarmReceiver called, save or update bonsai state")
+                CoroutineScope(Dispatchers.IO).launch {
+                    BonsaiStateManager().saveBonsaiState(context)
+                }
+            } else {
+                Log.e(TAG, "Context is null")
+            }
+        } else {
+            Log.e(TAG, "Unknown intent action ${intent.action}")
         }
     }
 }